+++ activerecord/test/cases/errors_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require "cases/helper"
-require "active_record/errors"
+require "active_record/exceptions"
 
 class ErrorsTest < ActiveRecord::TestCase
   def test_can_be_instantiated_with_no_args